Revert "Added Database logic functions"
