--- backend/src/config/corsConfig.js
@@ -1,18 +1,0 @@
-const cors = require('cors');
-
-const allowedOrigins = ['http://localhost:4000']; // Add your allowed origins here
-
-const corsOptions = {
-  origin: function (origin, callback) {
-    if (!origin || allowedOrigins.indexOf(origin) !== -1) {
-      callback(null, true);
-    } else {
-      callback(new Error('Not allowed by CORS'));
-    }
-  },
-  methods: ['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS'],
-  allowedHeaders: ['Content-Type', 'Authorization'],
-  credentials: true,
-};
-
-module.exports = cors(corsOptions);
