+++ backend/src/config/createTables.js
@@ -1,6 +1,6 @@
 const pool = require('./db');
 
-
+// Query to check if users table exists
 const checkUsersTableQuery = `
   SELECT EXISTS (
     SELECT FROM information_schema.tables 
@@ -8,15 +8,8 @@ const checkUsersTableQuery = `
     AND table_name = 'users'
   );
 `;
-const checkTransactionsTableQuery = `
-  SELECT EXISTS (
-    SELECT FROM information_schema.tables 
-    WHERE table_schema = 'public' 
-    AND table_name = 'transactions'
-  );
-`;
 
-
+// Query to create users table
 const createUserTableQuery = `
   CREATE TABLE IF NOT EXISTS users (
     id SERIAL PRIMARY KEY,
@@ -27,21 +20,6 @@ const createUserTableQuery = `
 
   );
 `;
-
-const createTransactionTableQuery = `
-  CREATE TABLE IF NOT EXISTS transactions (\
-    id SERIAL PRIMARY KEY,
-    user_id INTEGER NOT NULL,
-    title VARCHAR(255) NOT NULL , 
-    amount NUMERIC(10,2) NOT NULL,
-    description TEXT , 
-    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE
-
-  )
-
-
-`
 
 async function createTables() {
   try {
@@ -54,19 +32,12 @@ async function createTables() {
     } else {
       console.log('Users table already exists');
     }
-
-    // Check if transactions table exists
-    const { rows: transactionsTableExists } = await pool.query(checkTransactionsTableQuery);
-    if (!transactionsTableExists[0].exists) {
-      // Create transactions table
-      await pool.query(createTransactionTableQuery);
-      console.log('Transactions table created successfully');
-    } else {
-      console.log('Transactions table already exists');
-    }
   } catch (error) {
     console.error('Error creating tables:', error);
-  } 
+  } finally {
+  
+    pool.end();
+  }
 }
 
 module.exports = createTables;
