+++ backend/src/config/db.js
@@ -3,7 +3,6 @@ const path = require('path');
 const fs = require('fs');
 require('dotenv').config(); // Load environment variables from .env file
 
-
 const caCertPath = path.resolve(__dirname, process.env.CA_CERT_PATH);
 const pool = new Pool({
   user: process.env.PG_USER,
