+++ backend/src/routes/userRoutes.js
@@ -1,8 +1,23 @@
-
-const { checkEmailExists, checkPasswordCorrect, createUser  , getUserID, getBalance} = require('../services/userService');
+const express = require('express');
+const { checkEmailExists, checkPasswordCorrect, createUser  , getUserID} = require('../services/userService');
+const { isAuthenticated } = require('../middleware/autMiddleware'); 
+const router = express.Router();
 const jwt = require('jsonwebtoken');
 
-exports.login = async (req, res) => {
+router.post('/checkEmailExists', async (req, res) => {
+    const email = req.body.email;
+    const emailExists = await checkEmailExists(email);
+    res.send({ emailExists });
+});
+
+router.post('/checkPasswordCorrect', async (req, res) => {
+    const email = req.body.email;
+    const password = req.body.password;
+    const passwordCorrect = await checkPasswordCorrect(email, password);
+    res.send({ passwordCorrect });
+});
+
+router.post('/login', async (req, res) => {
     const { email, password } = req.body;
 
     try {
@@ -27,9 +42,9 @@ exports.login = async (req, res) => {
         console.error('Error in login:', error);
         res.status(500).send({ success: false, error: error.message });
     }
-}
-  
-  exports.signup = async (req, res) => {
+});
+
+router.post('/signup', async (req, res) => {
     const { email, password } = req.body;
 
     try {
@@ -56,15 +71,17 @@ exports.login = async (req, res) => {
         console.error('Error in signup:', error);
         res.status(500).send({ success: false, error: error.message });
     }
-}
-  
-  exports.balance = async (req, res) => {
-    const userID = req.user.id;
+});
+
+
+
+router.get('/logout', async (req, res) => {
     try {
-        const result = await getBalance(userID);
-        res.send({ result: result });
+        req.session.destroy();
+        res.send({ success: true });
     } catch (error) {
-        console.error('Error getting balance:', error);
         res.status(500).send({ success: false, error: error.message });
     }
-}
+});
+
+module.exports = router;
