+++ backend/src/services/userService.js
@@ -1,7 +1,6 @@
 const { system } = require('nodemon/lib/config');
 const pool = require('../config/db');
 const { hashPassword, comparePasswords } = require('./securePassword');
-const { use } = require('../routes/users');
 
 
 const getUserID = async (email) => {
@@ -71,69 +70,7 @@ const createUser = async (email, password) => {
         throw error;
     }
 }
-const makeTransaction = async ( userID , amount , title , description) => {
-    try {
-        const addTransactionQuery = {
-            text: 'INSERT INTO transactions (user_id , amount , title , description) VALUES ($1, $2, $3, $4)'
-                    ,
-            values : [ userID, amount , title , description ]
-        }
-        const changeBalanceQuery = {
-            text :  'UPDATE users SET balance = balance + $1 WHERE id = $2',
-            values: [amount , userID]
-        }
-        try {
-            const result = await pool.query(addTransactionQuery);
-            const result2 = await pool.query(changeBalanceQuery);
-            console.log(`succesfuly made transaction user_id : ${userID} , amount : ${amount} , title : ${title} , description : ${description}`)
-        } catch(error){
-            console.error('Error updating balance:', error);
-        }
-    } catch(error){
-        console.error("problem with decrease balance function",error);
-    }
-}
-const getBalance = async (userID) => {
-    try {
-        const query  = {
-            text: 'SELECT balance FROM users WHERE id = $1' ,
-            values : [userID]
-        }
-        try {
-            const result = await pool.query(query);
-            return result.rows[0];
-        } catch (error){
-            console.error ("problem when fetching balance" , error);
-        }
-    }catch{
-        console.error("problem before query", error);
-    }
-}
-const getUserTransactionsByPage = async(userID , pageNumber) => {
 
-    const pageSize = 10;
-    const offset = (pageNumber - 1) * pageSize;
-    try {
-        const query = {
-            text: `
-                SELECT * FROM transactions 
-                WHERE user_id = $1 
-                ORDER BY created_at ASC 
-                LIMIT $2 
-                OFFSET $3
-            `,
-            values: [userID, pageSize, offset]
-        };
-        try {
-            const result = await pool.query(query);
-            return result.rows;
-        } catch{
-            console.error("error when getting transactions" , error);
-        }
-    }catch {
-        console.error("error in function start" , error);
-    }
-}
 
 
 
@@ -143,6 +80,5 @@ module.exports = {
     checkEmailExists,
     checkPasswordCorrect,
     createUser, 
-    getUserID,
-    getBalance,
+    getUserID
 };
